+++ server/package.json
@@ -44,13 +44,13 @@
         "tslint": "^5.11.0"
     },
     "devDependencies": {
-        "@types/node": "^10.12.0",
         "@types/cookie-session": "^2.0.36",
         "@types/express": "^4.16.0",
         "@types/express-session": "^1.15.11",
+        "@types/node": "^10.12.0",
         "@types/passport": "^0.4.6",
         "@types/passport-google-oauth2": "^0.1.2",
-        "apollo-server-express": "^2.1.0",
+        "apollo-server-express": "^2.4.8",
         "concurrently": "^4.0.1",
         "ts-node": "^7.0.1",
         "typescript": "^2.6.1"
